remove old serializer file
